+++ app/helpers/UtilsHelper.php
@@ -775,3 +775,56 @@ function __()
     $string = array_shift($args);
     return $l->translate($string, $args);
 }
+
+/*
+ * @desc Get the browser name from a user agent
+ */
+function getBrowser(string $userAgent): ?string
+{
+    $t = strtolower($userAgent);
+    $t = ' ' . $t;
+
+    if     (strpos($t, 'opera'  )) return 'Opera';
+    elseif (strpos($t, 'edge'   )) return 'Edge';
+    elseif (strpos($t, 'chrome' )) return 'Chrome';
+    elseif (strpos($t, 'safari' )) return 'Safari';
+    elseif (strpos($t, 'firefox')) return 'Firefox';
+}
+
+/*
+ * @desc Get the platform from the user agent
+ */
+function getPlatform(string $userAgent): ?string
+{
+	$oses =  [
+		'/windows nt 10/i'      =>  'Windows 10',
+		'/windows nt 6.3/i'     =>  'Windows 8.1',
+		'/windows nt 6.2/i'     =>  'Windows 8',
+		'/windows nt 6.1/i'     =>  'Windows 7',
+		'/windows nt 6.0/i'     =>  'Windows Vista',
+		'/windows nt 5.2/i'     =>  'Windows Server 2003/XP x64',
+		'/windows nt 5.1/i'     =>  'Windows XP',
+		'/windows xp/i'         =>  'Windows XP',
+		'/windows nt 5.0/i'     =>  'Windows 2000',
+		'/windows me/i'         =>  'Windows ME',
+		'/win98/i'              =>  'Windows 98',
+		'/win95/i'              =>  'Windows 95',
+		'/win16/i'              =>  'Windows 3.11',
+		'/macintosh|mac os x/i' =>  'Mac OS X',
+		'/mac_powerpc/i'        =>  'Mac OS 9',
+		'/linux/i'              =>  'Linux',
+		'/ubuntu/i'             =>  'Ubuntu',
+		'/iphone/i'             =>  'iPhone',
+		'/ipod/i'               =>  'iPod',
+		'/ipad/i'               =>  'iPad',
+		'/android/i'            =>  'Android',
+		'/blackberry/i'         =>  'BlackBerry',
+		'/webos/i'              =>  'Mobile'
+    ];
+
+	foreach ($oses as $regex => $value) {
+		if (preg_match($regex, $userAgent)) {
+            return $value;
+		}
+	}
+}