+++ app/widgets/Presence/presence.js
@@ -4,10 +4,6 @@ var Presence = {
         localStorage.removeItem('quickLogin');
         localStorage.removeItem('quickHost');
         localStorage.removeItem('quickKey');
-    },
-    setFirebaseToken : function(token) {
-        Presence_ajaxSetFireBaseToken(token);
-        Android.clearFirebaseToken();
     }
 }